Quite a few changes:
added a reference to Moq and NUnit. Mock and Unit testing frameworks.
Changed the code to be mockable with Moq

Removed unnecessary exception.
Turned IDataStore into 3 sub Interfaces. No longer are we breaking the
Seperation of Concerns principle.
IJSONService, IApiService, IDataBroker was added.

Added a similiar structure to the Whmcs.Test project, to reflect the
structure in the Whmcs.Net project